fix race condition und speedups
there was a race condition when a process died, which still had children
attached to it. this caused some nasty memory corruption. the tree now moves
the childs eighter to the parent of the dead child or to init.
if init decides to move the child around, the internal tree gets fixed on the
next full update.

added changed filter to list_process which increases the scheduler code. on this
machine a full update takes now ~ 0.05 seconds, where parsing the /proc filesystem
is the slowest. There is still some optimization possible there, as we are not using
good openproc flags yet.

added trace log level printing really really verbose stuff like netlink messages

small fixes