die swap of death die
implement timeout of flags
add lua api to calculate timeout flags

add protector rule for detecting memory pressure
it detects memory pressure when caches and free ram drops below a
configureable percent of ram
it selects the target of isolation based on the rss size and caps
their limit of allowed physical ram with of the size of the detection
reduced with some percent.
swapping itself is not the only problem. if the filesystem caches drop
to low, the system becomes irresponsible because every read to every file
requires io traffic.
it sucessfull protects agains the memleak test program and compiling pypy
works great.