changed scheduling to be group based
use process group to be the identifier for groups and not session anymore.

this has advantage as this value is supposed to be used as such.
the process group can be set from the outside (added lua api for that).
spawned processes from shells for example get a now process group, so
the shell will be more responsive then the program executing, allowing
signals to be passed better.