implement fake group system
turns out that even as root it is not possible to issue a setgrpid on
a process, not even attached with ptrace as a fake parent.
added a workaround for that, allowing process groups to be set anytime.

add a kde filter and cleanup function that fixes the mess kde does at launching
jobs. all of them have the same session and task group, which a program started
by a desktop env should not have