add io group optimizer
group_isolation is a bad thing on very low load, but helps a lot
when the harddisk is very busy to give equal shares to the disk
even on random io. we monitor the weighted diskstat and if a threshold
is reached in the window size, group_isolation is enabled.
this causes a better responsive system on heavy io, but still allows
optimisation by the io scheduler on low io.