fix delay code/netlink code.
fix a grave bug in the new delay code. the parent of the fork message
is of course, not the parent of the new process. it is the process
who called fork and therefore, the new processes parent is parent->parent.

not enabled by default as the gnomefixer and kdefixer need some changes first.