add api call for number of running processes
get_number_of_processes() return an up to date number of running processes.
It's really sick to count the number of of directories that look like processes,
but the kernel simply does not export it in a sane way.

fix crash on subscribing to a new session on consolekit dbus errors