implement simple rule file filter
simplerules is a module that parses simple text files which allows
the admin to configure simple rules without the need to write
boilerplate lua code. The simplerules module is powerfull enough
to add simple flags to everything that can be matched based on
command line, executable path or basename of a process.