implement hard limits on memory usage
pressure groups now will get a hard limit on memory consumption in memory
and swap. This helps limiting run away processes to get killed before the oom
starts in most cases.

reincreased the free memory treshold to 10% as it seems to low with 5% on systems
with 1 and 2gb ram so the memory protector never starts. On systems with more then 4gb
ram the value could be decreased.