add single_task & game config
add scheduler config for games and single_task.
game scheduler tries to give most cpu to the game and system essentials like X server.
As it is even more harder to detect the running game, the current detection will fail
for games not started through the client and reside outside of standard game paths.
This detection will be improve later.

single_task tries to give all ressources to one task and his children.
single_task is only supposed to be used form command line through the client and should
be switched back after the program finished.