add video_playback config
experimental video_playback config tries to free as much cpu possible for video playback.
it uses cpuset to ensure the cpu running the videoplayer and essentials like
X server exclusivly on all except one core. this may help on systems that have
trouble running a movie if the cpu is on it's limit decoding the movie.