core: track real user sessions, reimplement consolekit agent as module
* u_session type renamed to USession
* sessions subsystem is implemented in usession.c
  - GList *U_session_list replaced with USession *U_sessions, which is
    pointer to the first USession instance.
  - u_sessions_*() family functions
* header for session tracking agents is usession-agent.h
* new USession structure members: is_valid, id, name, leader,
  consolekit_cookie
  - The "guint id" property is the actual session identification
    generated for every new session.

Consolekit agent was reimplemented into a module, other agents such
as logind may be easily added. Once the agent is registered to
the system, it's responsible for determining process sessions and
sessions leaders.

Currently, sessions are not exported to lua.

The code is heavily documented and is described by USession module
in doxygen generated documentation.