[UAssert] Add u_assert_process_dead (proc)
Print a warning if process exists in /proc/ directory and is not
a zombie.

In DEVELOP_MODE, this assertion is called everytime some process
is marked as dead, specifically everytime U_PROC_STATE(P, UPROC_DEAD)
macro is called. Assertion failure in this case is a symptom that
ulatencyd core is making false assumptions about /proc/ behaviour,
i.e. it falsely assumes some process is dead or it is a zombie but
instead it is still alive.