[core] Really make sure no thread goes through update_processes_run()
update_processes_run() skips every parsed process which is actually
a thread. This may not happen in full run but may happen if passed
proctab is polluted by TIDs. I already removed PROC_LOOSE_TASKS from
OPENPROC_FLAGS before, but this flag is only used in readproc2() which
we don't use, so it's useless to include or exclude it.