[core] proc_by_pid_with_retry(): Parse not yet known PIDs and TIDs
Parse a PID if not already known by the core and return updated
u_proc instance.
If passed PID is an unknown thread, parse its thread leader and returns
its u_proc instance.

NULL is returned only if a passed PID vanished from /proc/ or if it
is a thread which thread leader does no more exists.