[u_proc] Change and rename ENSURE_WHAT enum to be usable as a mask
It no more enumerates 'tags' (BASIC, CGROUP, ...) as a sequence of
integers but rather as binary flags. So the enumeration may be now
used as a mask of properties; and I renamed the enum to
U_PROC_PROPERTIES.

u_proc.ensured, which was structure of booleans, is replaced with
guint mask of U_PROC_PROPERTIES.

Structure u_proc_ensured is removed.