[mappings] Consider process w/ user.idle being active only if focused
This allows a process which creates X window to be labeled with
user.idle. If the window is focused, the process is scheduled as
active; if not, it is scheduled as user.idle.

Former behaviour was to consider as active all processes that are
in the focus stack, regardless their position and without taking
user.idle label into account.

This commit changes this behaviour:

A process which is tagged with user.idle flag is moved to usr_$$/idle
cpu cgroup unless it is focused, i.e. unless it's position in the
focus stack is 1, in which case it is moved to usr_$$/active;
similar for other subsystems.